requirements/base.txt
requests==2.26.0
factory_boy==3.2.1
pydantic-factories==1.5.4
